+++ src/main/java/com/cafromet/modelo/EspacioNatural.java
@@ -1,5 +1,5 @@
 package com.cafromet.modelo;
-// Generated 4 ene. 2021 19:38:01 by Hibernate Tools 5.4.21.Final
+// Generated 6 ene. 2021 20:21:45 by Hibernate Tools 5.4.21.Final
 
 import java.util.HashSet;
 import java.util.Set;
@@ -15,15 +15,15 @@ public class EspacioNatural implements java.io.Serializable {
 	private String tipo;
 	private String direccion;
 	private String categoria;
-	private Float latitud;
-	private Float longitud;
+	private Double latitud;
+	private Double longitud;
 	private Set municipio_EspacioNaturals = new HashSet(0);
 
 	public EspacioNatural() {
 	}
 
 	public EspacioNatural(String nombre, String descripcion, String tipo, String direccion, String categoria,
-			Float latitud, Float longitud, Set municipio_EspacioNaturals) {
+			Double latitud, Double longitud, Set municipio_EspacioNaturals) {
 		this.nombre = nombre;
 		this.descripcion = descripcion;
 		this.tipo = tipo;
@@ -82,19 +82,19 @@ public class EspacioNatural implements java.io.Serializable {
 		this.categoria = categoria;
 	}
 
-	public Float getLatitud() {
+	public Double getLatitud() {
 		return this.latitud;
 	}
 
-	public void setLatitud(Float latitud) {
+	public void setLatitud(Double latitud) {
 		this.latitud = latitud;
 	}
 
-	public Float getLongitud() {
+	public Double getLongitud() {
 		return this.longitud;
 	}
 
-	public void setLongitud(Float longitud) {
+	public void setLongitud(Double longitud) {
 		this.longitud = longitud;
 	}
